Schema-Intelligence-layer/backend/requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2-binary
schema-matching-toolkit
rank_bm25
groq
neo4j
pandas
sqlalchemy
networkx